change register user to Firestore location
